Split app registration boundary
